# public/swagger.json
{"swagger":"2.0","info":{"title":"SenseHome-Technologies API Hubs","version":"1.0.0","description":"The purpose of the API developed is to manage the hubs of a system, including devices, events, divisions and information management functions."},"paths":{"/api/device/add":{"post":{"tags":["Device"],"summary":"Add a device to a hub","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Add a device to a hub","required":true,"schema":{"$ref":"#/definitions/AddDevice"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/device"}}}}}},"/api/device/create":{"post":{"tags":["Device"],"summary":"Create a device","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Create a device","required":true,"schema":{"$ref":"#/definitions/CreateDevice"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/device"}}}}}},"/api/device/delete":{"delete":{"tags":["Device"],"summary":"Delete a device by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"The ID of the device to delete"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/device"}}}}}},"/api/device/edit":{"put":{"tags":["Device"],"summary":"Edit a device by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Edit a device by its ID","required":true,"schema":{"$ref":"#/definitions/EditDevice"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/device"}}}}}},"/api/device/getall":{"get":{"tags":["Device"],"summary":"Retrieve all devices of a hub by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"hubid","in":"path","required":true,"type":"string","description":"The ID of the hub to retrieve devices from"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/device"}}}}}},"/api/device/get":{"get":{"tags":["Device"],"summary":"Retrieve a device by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"The ID of the device to retrieve"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/device"}}}}}},"/api/event/create":{"post":{"tags":["Event"],"summary":"Create an event","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Create an event","required":true,"schema":{"$ref":"#/definitions/CreateEvent"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/event"}}}}}},"/api/event/delete":{"delete":{"tags":["Event"],"summary":"Delete an event by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"The ID of the event to delete"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/event"}}}}}},"/api/event/edit":{"put":{"tags":["Event"],"summary":"Update an event with the provided data","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Update an event with the provided data","required":true,"schema":{"$ref":"#/definitions/EditEvent"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/event"}}}}}},"/api/event/getall":{"get":{"tags":["Event"],"summary":"Retrieve all events of a hub by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"hubid","in":"path","required":true,"type":"string","description":"The ID of the hub to retrieve events from"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/event"}}}}}},"/api/event/get":{"get":{"tags":["Event"],"summary":"Retrieve an event by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"The ID of the event to retrieve"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/event"}}}}}},"/api/hub/add":{"post":{"tags":["Hub"],"summary":"Add a hub by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Add a hub by its ID","required":true,"schema":{"$ref":"#/definitions/AddHub"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/hub"}}}}}},"/api/hub/delete":{"delete":{"tags":["Hub"],"summary":"Delete a hub by its ID and type","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"The ID of the hub to delete"},{"name":"type","in":"path","required":true,"type":"string","description":"The type of the hub to delete"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/hub"}}}}}},"/api/hub/edit":{"put":{"tags":["Hub"],"summary":"Edit a hub by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Edit a hub by its ID","required":true,"schema":{"$ref":"#/definitions/EditHub"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/hub"}}}}}},"/api/hub/getall":{"get":{"tags":["Hub"],"summary":"Retrieve all hubs associated with the user's email","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"email","in":"path","required":true,"type":"string","description":"The email of the user to retrieve hubs for"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/hub"}}}}}},"/api/hub/get":{"get":{"tags":["Hub"],"summary":"Retrieve a hub by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"The ID of the hub to retrieve"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/hub"}}}}}},"/api/hub/login":{"post":{"tags":["Hub"],"summary":"Login a hub by its ID","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Login a hub by its ID","required":true,"schema":{"$ref":"#/definitions/LoginHub"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/hub"}}}}}},"/api//devicetypes/get":{"get":{"tags":["Types"],"summary":"Retrieve all available device types","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/response"}}}}}},"/api//eventtypes/get":{"get":{"tags":["Types"],"summary":"Retrieve all available event types","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/response"}}}}}},"/api//operators/get":{"get":{"tags":["Types"],"summary":"Retrieve all available operators","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/response"}}}}}},"/api//statments/get":{"get":{"tags":["Types"],"summary":"Retrieve all available statements","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object","items":{"$ref":"#/definitions/response"}}}}}}},"definitions":{"AddDevice":{"properties":{"hubid":{"type":"string","description":"The hubid to add the device to"}},"required":["hubid"]},"device":{"properties":{"status":{"type":"number","description":"The status of the device"},"message":{"type":"string","description":"The message of the device"}},"required":["status","message"]},"CreateDevice":{"properties":{"name":{"type":"string","description":"The name of the device"},"accesscode":{"type":"string","description":"The access code of the device"},"type":{"type":"number","description":"The type of the device"},"state":{"type":"number","description":"The state of the device"},"value":{"type":"number","description":"The value of the device"}},"required":["name","accesscode","type","state","value"]},"DeleteDevice":{"properties":{"id":{"type":"string","description":"The ID of the device to delete"}},"required":["id"]},"EditDevice":{"properties":{"id":{"type":"string","description":"The ID of the device to edit"},"name":{"type":"string","description":"The name of the device"},"accesscode":{"type":"string","description":"The access code of the device"},"type":{"type":"string","description":"The type of the device"},"state":{"type":"string","description":"The state of the device"},"value":{"type":"string","description":"The value of the device"},"favorite":{"type":"boolean","description":"The favorite status of the device"},"hubid":{"type":"string","description":"The ID of the hub of the device"},"divisionid":{"type":"string","description":"The ID of the division of the device"}},"required":["id","name","accesscode","type","state","value","favorite","hubid","divisionid"]},"GetAllDevices":{"properties":{"hubid":{"type":"string","description":"The ID of the hub to retrieve devices from"}},"required":["hubid"]},"GetDevice":{"properties":{"id":{"type":"string","description":"The ID of the device to retrieve"}},"required":["id"]},"CreateEvent":{"properties":{"name":{"type":"string","description":"The name of the event"},"type":{"type":"string","description":"The type of the event"},"state":{"type":"string","description":"The state of the event"},"hubid":{"type":"string","description":"The id of the hub to which the event belongs"},"eventtargets":{"type":"array","description":"The list of event targets","items":{"type":"Object"}},"schedules":{"type":"array","description":"The list of schedules","items":{"type":"Object"}}},"required":["name","type","state","hubid","eventtargets","schedules"]},"event":{"properties":{"status":{"type":"number","description":"The status of the event"},"message":{"type":"string","description":"The message of the event"}},"required":["status","message"]},"DeleteEvent":{"properties":{"id":{"type":"string","description":"The ID of the event to delete"}},"required":["id"]},"EditEvent":{"properties":{"id":{"type":"string","description":"The ID of the event to update"},"name":{"type":"string","description":"The new name of the event"},"type":{"type":"string","description":"The new type of the event"},"state":{"type":"string","description":"The new state of the event"},"hubid":{"type":"string","description":"The ID of the hub the event belongs to"},"eventtargets":{"type":"array","description":"The updated list of event targets","items":{"type":"Object"}},"schedules":{"type":"array","description":"The updated list of schedules","items":{"type":"Object"}}},"required":["id","name","type","state","hubid","eventtargets","schedules"]},"GetAllEvents":{"properties":{"hubid":{"type":"string","description":"The ID of the hub to retrieve events from"}},"required":["hubid"]},"GetEvent":{"properties":{"id":{"type":"string","description":"The ID of the event to retrieve"}},"required":["id"]},"AddHub":{"properties":{"id":{"type":"string","description":"The ID of the hub to add"},"email":{"type":"string","description":"The email to register the hub with"}},"required":["id","email"]},"hub":{"properties":{"status":{"type":"number","description":"The status of the hub"},"message":{"type":"string","description":"The message of the hub"},"token":{"type":"string","description":"The JWT token of the hub"}},"required":["status","message","token"]},"DeleteHub":{"properties":{"id":{"type":"string","description":"The ID of the hub to delete"},"type":{"type":"string","description":"The type of the hub to delete"}},"required":["id","type"]},"EditHub":{"properties":{"id":{"type":"string","description":"The ID of the hub to edit"},"name":{"type":"string","description":"The name of the hub to update"},"users":{"type":"array","description":"The users to update in the hub"}},"required":["id","name","users"]},"GetAllHubs":{"properties":{"email":{"type":"string","description":"The email of the user to retrieve hubs for"}},"required":["email"]},"GetHub":{"properties":{"id":{"type":"string","description":"The ID of the hub to retrieve"}},"required":["id"]},"LoginHub":{"properties":{"id":{"type":"string","description":"The ID of the hub to login"}},"required":["id"]},"response":{"properties":{"status":{"type":"number","description":"The status code of the response"},"message":{"type":"string","description":"The message of the response"},"data":{"type":"array","description":"The list of statements"}},"required":["status","message","data"]}}}